[node-api] extract send and watch tx.
